+++ core/src/Online/cliente.java
@@ -38,7 +38,7 @@ public class cliente {
     public static void enviarMensaje(String string){
           hiloC.enviarMensaje(string);
     }
-    public static int getCantidad(){
+    public static float getCantidad(){
         return hiloC.getCant();
     }
